# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Valuz Agent API
  version: "0.3.0"
  description: Valuz Agent API contract — single source of truth
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /v1/onboarding/example-project:
    post:
      summary: Create (or reuse) the onboarding example project
      operationId: createExampleProject
      tags: [onboarding]
      description: |
        One-shot endpoint called at the end of the onboarding flow.
        Creates ~/Valuz/示例项目, binds it as a project, and deploys
        the four official agents for the selected preset team.
        Fully idempotent — safe to call again if the user re-runs onboarding.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExampleProjectRequest"
      responses:
        "200":
          description: Project created or reused successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExampleProjectResponse"
        "422":
          description: Invalid team_id or project creation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/onboarding/assistant:
    post:
      summary: Create (or reuse) the general Valuz assistant
      operationId: createAssistant
      tags: [onboarding]
      description: |
        Creates (or reuses) the general 「Valuz 小助手」 in the user's agent
        library — backs TeamStep's「暂时不需要 team」choice. No project is
        created. Idempotent. Returns 422 if no model channel is configured.
      responses:
        "200":
          description: Assistant created or reused
          content:
            application/json:
              schema:
                type: object
                required: [agent_slug]
                properties:
                  agent_slug: {type: string}
        "422":
          description: No model channel configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...